+++ core/parsers/vpk_file.py
