+++ operations/advanced_particle_merger.py
@@ -1,7 +1,7 @@
 from collections import defaultdict
 from pathlib import Path
 from typing import Dict, List
-from parsers.pcf_file import PCFFile
+from core.parsers.pcf_file import PCFFile
 from operations.pcf_rebuild import (
     load_particle_system_map,
     get_pcf_element_names,
