+++ operations/file_processors.py
@@ -1,5 +1,5 @@
 from pathlib import Path
-from parsers.pcf_file import PCFFile
+from core.parsers.pcf_file import PCFFile
 from operations.pcf_compress import remove_duplicate_elements
