+++ operations/pcf_merge.py
@@ -1,4 +1,4 @@
-from parsers.pcf_file import PCFFile, PCFElement
+from core.parsers.pcf_file import PCFFile, PCFElement
 from core.constants import AttributeType
